Initial commit with v1.0.0
This project was made possible by the efforts of myself, Dillon Lloyd, Preethi Thunga, Skylar Marvel, and David Reif.

Briefly, myself, Dillon, and David conceived of the package structure. Myself and Dillon wrote the bulk of the package code.

Preethi and Skylar provided testing and feedback, code patches, and wrote the user vignettes.

Co-Authored-By: Dillon Lloyd <51928118+dtlloyd96@users.noreply.github.com>
Co-Authored-By: Preethi Thunga <pthunga@ncsu.edu>
Co-Authored-By: SkylarMarvel <57526164+SkylarMarvel@users.noreply.github.com>
Co-Authored-By: David Reif <42413358+davidmreif@users.noreply.github.com>